[fix] 07/08 history 없을 때 뒤로가 기 버튼 클릭 시 홈화면으로 라우팅 #85
