+++ app/components/header/HeaderBackButton.tsx
@@ -1,0 +1,46 @@
+import BackIcon from "@/public/icons/chevron-left.svg";
+import { useSearchKeywordStore } from "@/store/useSearchKeywordStore";
+import { usePathname, useRouter, useSearchParams } from "next/navigation";
+
+const HeaderBackButton = () => {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const pathname = usePathname();
+  const { keyword, setKeyword } = useSearchKeywordStore();
+
+  const handleBack = () => {
+    const page = searchParams.get("page");
+
+    if (
+      pathname.startsWith("/recruit/") &&
+      pathname !== "/recruit" &&
+      keyword
+    ) {
+      const query = new URLSearchParams();
+      query.set("keyword", keyword);
+      if (page) query.set("page", page);
+
+      router.push(`/recruit?${query.toString()}`);
+    } else if (keyword) {
+      setKeyword("");
+      router.replace("/recruit");
+    } else {
+      if (
+        typeof window !== "undefined" &&
+        document.referrer &&
+        new URL(document.referrer).origin === window.location.origin
+      ) {
+        router.back();
+      } else {
+        router.push("/");
+      }
+    }
+  };
+  return (
+    <button type="button" onClick={handleBack}>
+      <BackIcon width="24px" height="24px" color="#171717" />
+    </button>
+  );
+};
+
+export default HeaderBackButton;
