+++ app/globals.css
@@ -89,6 +89,10 @@ body {
     @apply text-[14px] font-normal leading-[130%] tracking-[-0.02em];
   }
 
+  .text-body-4 {
+    @apply text-[14px] font-normal leading-[160%] tracking-[-0.02em];
+  }
+
   .text-caption-1 {
     @apply text-[14px] font-bold leading-[130%] tracking-[-0.02em];
   }
